Removed SearchIndex __call__ function - Handle search query parameter processing and execution in `CloudantDatabase.get_search_result` - Moved search query test cases into CloudantDatabaseTests and renamed tests after get_search_result function - Search index function is now a required argument in SearchIndex init and unit tests updated to reflect this change - Moved search_url function into common_util module
