+++ tests/unit/unit_t_db_base.py
@@ -235,11 +235,11 @@ class UnitTestDbBase(unittest.TestCase):
         Create a design document with search indexes for use
         with search query tests.
         """
-        self.ddoc = DesignDocument(self.db, 'ddoc001')
-        self.ddoc['indexes'] = {'searchindex001': {
+        self.search_ddoc = DesignDocument(self.db, 'searchddoc001')
+        self.search_ddoc['indexes'] = {'searchindex001': {
                 'index': 'function (doc) {\n  index("default", doc._id); \n '
                 'if (doc.name) {\n index("name", doc.name, {"store": true}); \n} '
                 'if (doc.age) {\n index("age", doc.age, {"facet": true}); \n} \n} '
             }
         }
-        self.ddoc.save()
+        self.search_ddoc.save()
